+++ plugins/prism.php
@@ -34,7 +34,7 @@ import '<?php echo $this->editorRoot; ?>/prism/languages/sql.js';
 adminerHighlighter = els => els.forEach(el => {
 	const mode = (
 		/jush-js/.test(el.className) ? 'json' : (
-		/jush-\w*sql/.test(el.className) ? 'sql' : (
+		/jush-(\w*sql|oracle|clickhouse|firebird)/.test(el.className) ? 'sql' : (
 		''
 	)));
 	if (mode) {