+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 - Elasticsearch: Make it work with Elasticsearch 8
 - CSS: Hide menu on mobile
 - CSS: Invert icons in dark mode
+- Plugins: Allow changing CSP by more plugins
 
 ## Adminer 5.1.0 (released 2025-03-24)
 - Display collation at table structure if different from table