+++ CHANGELOG.md
@@ -1,4 +1,4 @@
-## Adminer dev
+## Adminer 5.1.1 (released 2025-04-02)
 - Export: Fix tar (regression from 5.0.3)
 - Select: Allow ordering by COUNT(*) (bug #966, regression from 5.0.2)
 - Optimize retrieving columns for schema